Move cats-related code to `memeid-cats` module
