+++ cats/src/test/scala/memeid/cats/UUIDLaws.scala
@@ -1,8 +1,10 @@
-package memeid
+package memeid.cats
 
 import cats.instances.option._
 import cats.kernel.laws.discipline.{HashTests, LowerBoundedTests, OrderTests, UpperBoundedTests}
 
+import memeid.UUID
+import memeid.cats.instances._
 import org.specs2.mutable.Specification
 import org.typelevel.discipline.specs2.mutable.Discipline
