+++ core/src/test/scala/memeid/UUIDSpec.scala
@@ -1,10 +1,6 @@
 package memeid
 
 import java.nio.ByteBuffer
-import java.util.{UUID => JUUID}
-
-import cats._
-import cats.instances.uuid._
 
 import memeid.digest.Digestible
 import org.scalacheck.Gen
@@ -26,25 +22,6 @@ class UUIDSpec extends Specification with ScalaCheck {
       }
     }.setGen(Gen.alphaNumStr)
 
-  }
-
-  "Order[UUID]" should {
-
-    "compare using unsigned comparison" in prop { lsb: Long =>
-      // specifically chosen since they will not compare correctly unless using unsigned comparison
-      val uuid1 = UUID.from(0x20000000.toLong, lsb)
-      val uuid2 = UUID.from(0xE0000000.toLong, lsb)
-
-      (Order[UUID].compare(uuid1, uuid2) must be equalTo -1) and
-        (Order[JUUID].compare(uuid1.juuid, uuid2.juuid) must be equalTo 1)
-    }
-
-  }
-
-  "Hash[UUID] returns hash code consistent with java.util.UUID" in prop { (msb: Long, lsb: Long) =>
-    val uuid = UUID.from(msb, lsb)
-
-    Hash[UUID].hash(uuid) must be equalTo Hash[JUUID].hash(uuid.juuid)
   }
 
   "UUID.as" should {
