+++ core/src/test/scala/memeid/V3Spec.scala
@@ -1,10 +1,9 @@
 package memeid
 
 import org.specs2.ScalaCheck
-import org.specs2.matcher.IOMatchers
 import org.specs2.mutable.Specification
 
-class V3Spec extends Specification with ScalaCheck with IOMatchers {
+class V3Spec extends Specification with ScalaCheck {
 
   "V3 constructor" should {
